+++ R/plot_boxplot.R
@@ -13,6 +13,7 @@
 #' @param nrow number of rows per page
 #' @param ncol number of columns per page
 #' @param parallel enable parallel? Default is \code{FALSE}.
+#' @param plotly if \code{TRUE}, convert to interactive plotly object (requires the \pkg{plotly} package). Default is \code{FALSE}.
 #' @return invisibly return the named list of ggplot objects
 #' @keywords plot_boxplot
 #' @import data.table
@@ -41,7 +42,7 @@ plot_boxplot <- function(data, by,
                          title = NULL,
                          ggtheme = theme_gray(), theme_config = list(),
                          nrow = 3L, ncol = 4L,
-                         parallel = FALSE) {
+                         parallel = FALSE, plotly = FALSE) {
   ## Declare variable first to pass R CMD check
   variable <- by_f <- value <- NULL
   ## Check if input is data.table
@@ -85,6 +86,7 @@ plot_boxplot <- function(data, by,
     title = title,
     ggtheme = ggtheme,
     theme_config = theme_config,
+    plotly = plotly,
     facet_wrap_args = list(
       "facet" = ~ variable,
       "nrow" = nrow,