+++ R/plot_qq.R
@@ -12,6 +12,7 @@
 #' @param nrow number of rows per page. Default is 3.
 #' @param ncol number of columns per page. Default is 3.
 #' @param parallel enable parallel? Default is \code{FALSE}.
+#' @param plotly if \code{TRUE}, convert to interactive plotly object (requires the \pkg{plotly} package). Default is \code{FALSE}.
 #' @return invisibly return the named list of ggplot objects
 #' @keywords plot_qq
 #' @import data.table
@@ -32,7 +33,7 @@ plot_qq <- function(data, by = NULL, sampled_rows = nrow(data),
                     title = NULL,
                     ggtheme = theme_gray(), theme_config = list(),
                     nrow = 3L, ncol = 3L,
-                    parallel = FALSE) {
+                    parallel = FALSE, plotly = FALSE) {
   ## Declare variable first to pass R CMD check
   variable <- value <- group <- NULL
   ## Check if input is data.table
@@ -86,6 +87,7 @@ plot_qq <- function(data, by = NULL, sampled_rows = nrow(data),
     title = title,
     ggtheme = ggtheme,
     theme_config = theme_config,
+    plotly = plotly,
     facet_wrap_args = list(
       "facet" = ~ variable,
       "nrow" = nrow,